Fix broken due dates
The root cause of the bug is that Todoist returns `null` for the `due_date_utc`
field when no due date is set. Go's json decoder simply ignores null values,
so the DueDateUtc field wasn't being set correctly.

It was fixed by making DueDateUtc a string pointer. Go's json decoder does set
a string pointer to `nil` if the json value is `null`.

This should fix #68.