Add a command to show today's tasks only
Using the filter (-f) to get the list of today's items works, but it is
a bit time consuming to type it out every time. While creating an alias
for this is easy, I would like to see it as a default command on this
cli.

If this feature has a chance of being merged, please respond with some
feedback. I will accomodate suggestions and write test cases to make it
a whole.